Re-factor Ash Message Center code part 4/4 Extracts WebNotificationList management code to MessageCenter class.
After this the bulk of code can be moved to Views!

Updated the diagram:
https://docs.google.com/drawings/d/1B6ZxXObmrDFzgWEHmkP_Ppt2BFbOL9kXzakOmY0sXe0/edit


BUG=150872




Review URL: https://chromiumcodereview.appspot.com/11189099

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163409 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3632e69cbe4a559644c84b3bba5d0bc0fb65723d

+++ ash/system/web_notification/web_notification_list.h
@@ -44,6 +44,7 @@ class ASH_EXPORT WebNotificationList {
   explicit WebNotificationList(Delegate* delegate);
   virtual ~WebNotificationList();
 
+  // Affects whether or not a message has been "read".
   void SetMessageCenterVisible(bool visible);
 
   void AddNotification(const std::string& id,
@@ -84,7 +85,7 @@ class ASH_EXPORT WebNotificationList {
   void MarkPopupsAsShown();
 
   const Notifications& notifications() const { return notifications_; }
-  int unread_count() const { return unread_count_; }
+  size_t unread_count() const { return unread_count_; }
 
   static const size_t kMaxVisiblePopupNotifications;
   static const size_t kMaxVisibleMessageCenterNotifications;
@@ -106,7 +107,7 @@ class ASH_EXPORT WebNotificationList {
   Delegate* delegate_;
   Notifications notifications_;
   bool message_center_visible_;
-  int unread_count_;
+  size_t unread_count_;
 
   DISALLOW_COPY_AND_ASSIGN(WebNotificationList);
 };
